refactor(hooks): 封装 useBoundary hook 用于边界样式管理
- 新增 useBoundary hook，提供可复用的边界样式管理逻辑
- 支持延迟清除边界样式，默认延迟 300ms
- 在 Dropdown 组件中使用新的 useBoundary hook 替换原有实现
- 优化 AbsoluteList 的 props 传递，使用更简洁的逻辑与运算
