+++ packages/hooks/src/index.ts
@@ -30,6 +30,7 @@ export * from './common/use-filter';
 export * from './common/use-scrollbar-width';
 export * from './common/use-component-memo';
 export * from './common/use-collapse-animation';
+export * from './common/use-boundary';
 
 //components
 export * from './components/use-input';
